pyproject.toml
[project]
name = "stock-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.0.4",
    "google-generativeai>=0.8.5",
    "streamlit>=1.46.0",
]
